// core/codeapps/workflow-console/src/generated/models/Faf001_bptasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Faf001_bptasksfaf001_outcome = {
  324010000: 'Approved',
  324010001: 'Rejected',
  324010002: 'Needs More Info',
  324010003: 'Escalated',
  324010004: 'Completed',
  324010005: 'No Response'
} as const;
export type Faf001_bptasksfaf001_outcome = keyof typeof Faf001_bptasksfaf001_outcome;
export const Faf001_bptasksfaf001_status = {
  324010000: 'Not Started',
  324010001: 'Assigned',
  324010002: 'In Progress',
  324010003: 'Waiting',
  324010004: 'Completed',
  324010005: 'Cancelled',
  324010006: 'Expired'
} as const;
export type Faf001_bptasksfaf001_status = keyof typeof Faf001_bptasksfaf001_status;
export const Faf001_bptasksfaf001_tasktype = {
  324010000: 'Approval',
  324010001: 'Data Review',
  324010002: 'Data Entry',
  324010003: 'Custom'
} as const;
export type Faf001_bptasksfaf001_tasktype = keyof typeof Faf001_bptasksfaf001_tasktype;
export const Faf001_bptasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Faf001_bptasksstatecode = keyof typeof Faf001_bptasksstatecode;
export const Faf001_bptasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Faf001_bptasksstatuscode = keyof typeof Faf001_bptasksstatuscode;

export interface Faf001_bptasksBase {
  faf001_assignedon?: string;
  "faf001_assignedteam@odata.bind"?: string;
  "faf001_assignedto@odata.bind"?: string;
  "faf001_bpinstanceid@odata.bind": string;
  "faf001_bpstepid@odata.bind"?: string;
  faf001_bptaskid: string;
  faf001_comments?: string;
  faf001_completedon?: string;
  faf001_duedate?: string;
  faf001_inputdata?: string;
  faf001_outcome?: Faf001_bptasksfaf001_outcome;
  "faf001_requiredrole@odata.bind"?: string;
  faf001_status: Faf001_bptasksfaf001_status;
  faf001_taskname: string;
  faf001_tasktype?: Faf001_bptasksfaf001_tasktype;
  faf001_widgetid?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Faf001_bptasksstatecode;
  statuscode?: Faf001_bptasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Faf001_bptasks extends Faf001_bptasksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  faf001_assignedteamname?: string;
  faf001_assignedteamyominame?: string;
  faf001_assignedtoname?: string;
  faf001_assignedtoyominame?: string;
  faf001_bpinstanceidname?: string;
  faf001_bpstepidname?: string;
  faf001_outcomename?: string;
  faf001_requiredrolename?: string;
  faf001_statusname?: string;
  faf001_tasktypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  faf001_assignedteam?: object;
  _faf001_assignedteam_value?: string;
  faf001_assignedto?: object;
  _faf001_assignedto_value?: string;
  faf001_bpinstanceid?: object;
  _faf001_bpinstanceid_value?: string;
  faf001_bpstepid?: object;
  _faf001_bpstepid_value?: string;
  faf001_requiredrole?: object;
  _faf001_requiredrole_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}